base_project_django/pyproject.toml
[tool.poetry]
name = "base-project-django"
version = "0.1.0"
description = "base django template"
authors = ["nleea <neldecas12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-cors-headers = "^4.6.0"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg = "^3.2.3"


[tool.poetry.group.dev.dependencies]
drf-spectacular = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
